se cambio el tipo de dato para la cedula de personas y se elimina la validacion de correo unico para empresa

+++ database/migrations/2022_04_22_170418_create_personas_table.php
@@ -21,7 +21,7 @@ class CreatePersonasTable extends Migration
             $table->text('segundo_apellido_persona')->nullable();
             $table->text('genero_persona');
             $table->text('tipo_iden_persona');
-            $table->integer('cedula_persona')->unique()->nullable();
+            $table->bigInteger('cedula_persona')->unique()->nullable();
             $table->string('correo_persona', 255)->nullable();
             $table->string('telefono_persona', 15)->nullable();
             $table->string('lider_ava', 50)->nullable();
